+++ authentik/flows/exceptions.py
@@ -2,7 +2,7 @@
 
 from django.utils.translation import gettext_lazy as _
 
-from authentik.lib.tracing import TracingIgnoredException
+from authentik.lib.tracing.exceptions import TracingIgnoredException
 from authentik.policies.types import PolicyResult
