+++ authentik/policies/exceptions.py
@@ -1,6 +1,6 @@
 """policy exceptions"""
 
-from authentik.lib.tracing import TracingIgnoredException
+from authentik.lib.tracing.exceptions import TracingIgnoredException
 
 
 class PolicyEngineException(TracingIgnoredException):
