+++ authentik/sources/oauth/exceptions.py
@@ -1,6 +1,6 @@
 """OAuth Source Exception"""
 
-from authentik.lib.tracing import TracingIgnoredException
+from authentik.lib.tracing.exceptions import TracingIgnoredException
 
 
 class OAuthSourceException(TracingIgnoredException):
